remove duplicate code examples because of merge
