--- DemoProject/Tests/a_test.py
@@ -1,9 +1,0 @@
-from DemoProject.module2.code1 import *
-
-
-def test1():
-    assert (func1(5) == 6)
-
-
-def test2():
-    assert (func2([3, 3], 1) == 3)
