--- code_examples/TestsExamples/test_b.py
@@ -1,15 +1,0 @@
-# from code2.code2 import *
-# from code2.code1 import *
-#
-#
-# def test1():
-#     assert (func4(5) == 6)
-#
-#
-# # def test2():
-# #     assert (func5([3, 3], 1) == 3)
-#
-#
-# def test_person():
-#     p = Person("bob", 20)
-#     assert p.get_name() == "bob"
