--- code_examples/code2/code1.py
@@ -1,7 +1,0 @@
-class Person:
-    def __init__(self, name, age):
-        self.name = name
-        self.age = age
-
-    def get_name(self):
-        return self.name
