--- code_examples/code2/code2.py
@@ -1,1 +1,0 @@
-def func4(x): return x + 1
